Created mock/test table and resource page to see if implementation works
